Check if table exists before attempting to drop

+++ db_setup.py
@@ -3,24 +3,28 @@ import sqlite3
 def connect_db():
     """Connects to or creates a db and returns a cursor"""
     conn = sqlite3.connect('word_problems.db')
-    c = conn.cursor()
-    return c
+    return conn.cursor()
 
 def make_table(c, table_name, columns):
     c.execute('CREATE TABLE {} {}'.format(table_name, columns))
 
 def delete_table(c, table_name):
-    c.execute('DROP TABLE {}'.format(table_name))
+    tables = c.execute('''SELECT name FROM sqlite_master 
+                          WHERE type='table' AND name=?;''', (table_name,))
+    tables = tables.fetchall()
+    if tables:
+        c.execute('DROP TABLE {}'.format(table_name))
 
 def fill_table(c, file_name):
-    """Reads a .tsv file and fills a db table where the 1st and 4th columns are int"""
+    """Reads .tsv file, populates db table where columns 1 & 4 are int"""
     with open(file_name, 'rU') as f:
         for record in f.read().split('\n')[1:]:
             if record:
                 record = record.split('\t')
                 record[0] = float(record[0])
                 record[3] = float(record[3])
-                c.execute('''INSERT INTO problems VALUES (?,?,?,?,?,?,?,?,?)''', record)
+                c.execute('''INSERT INTO problems VALUES (?,?,?,?,?,?,?,?,?)''', 
+                          record)
 
 def print_table(c, table_name):
     for row in c.execute('SELECT * FROM {}'.format(table_name)):
